[ADVAPP-1168]: Improve and add tests for emails being sent with different demo mode configurations
